# docs/specs/api/openapi.yaml
openapi: 3.1.0
info:
  title: GabFlow API
  version: 0.4.0
  description: API principal para gestão de gabinetes municipais.
servers:
  - url: https://api.gabflow.example/api/v1
security:
  - cookieAuth: []
paths:
  /solicitacoes:
    get:
      summary: Lista solicitações
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: categoriaId
          schema:
            type: string
            format: uuid
        - in: query
          name: bairroId
          schema:
            type: string
            format: uuid
        - in: query
          name: page
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Página de solicitações
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SolicitacaoPage"
    post:
      summary: Cria uma solicitação
      operationId: criarSolicitacao
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriarSolicitacaoRequest"
      responses:
        "201":
          description: Solicitação criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Solicitacao"
        "422":
          $ref: "#/components/responses/ValidationError"
  /solicitacoes/{id}:
    get:
      summary: Obtém uma solicitação
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Solicitação
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Solicitacao"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      summary: Atualiza campos permitidos
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              type: object
      responses:
        "200":
          description: Solicitação atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Solicitacao"
        "422":
          $ref: "#/components/responses/ValidationError"
  /solicitacoes/{id}/interacoes:
    post:
      summary: Registra uma interação
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriarInteracaoRequest"
      responses:
        "201":
          description: Interação criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Solicitacao"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token_cookie
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
  responses:
    NotFound:
      description: Recurso não encontrado
    ValidationError:
      description: Erro de validação
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    CriarSolicitacaoRequest:
      type: object
      required: [origem, descricao]
      properties:
        cidadaoId:
          type: string
          format: uuid
        origem:
          type: string
          enum: [PRESENCIAL, TELEFONE, WHATSAPP, EMAIL, FORMULARIO, REDE_SOCIAL, VISITA]
        titulo:
          type: string
          maxLength: 180
        descricao:
          type: string
          minLength: 3
        endereco:
          type: string
          maxLength: 500
        categoria:
          type: string
          maxLength: 100
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        anexos:
          type: array
          items:
            type: string
            format: uuid
    Solicitacao:
      type: object
      required: [id, protocolo, origem, descricao, status, criadaEm]
      properties:
        id:
          type: string
          format: uuid
        protocolo:
          type: string
        origem:
          type: string
        titulo:
          type: string
        descricao:
          type: string
        status:
          type: string
          enum: [NOVA, TRIAGEM, EM_ATENDIMENTO, AGUARDANDO_ORGAO, AGUARDANDO_CIDADAO, RESOLVIDA, ENCERRADA, CANCELADA]
        prioridade:
          type: string
          enum: [BAIXA, MEDIA, ALTA, CRITICA]
        categoriaId:
          type: string
          format: uuid
        categoria:
          type: string
        subcategoriaId:
          type: string
          format: uuid
        responsavelId:
          type: string
          format: uuid
        endereco:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        motivoEncerramento:
          type: string
        evidenciaEncerramento:
          type: string
        encerradaEm:
          type: string
          format: date-time
        interacoes:
          type: array
          items:
            $ref: "#/components/schemas/Interacao"
        historico:
          type: array
          items:
            $ref: "#/components/schemas/HistoricoSolicitacao"
        criadaEm:
          type: string
          format: date-time
        atualizadaEm:
          type: string
          format: date-time
    SolicitacaoPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/Solicitacao"
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
    CriarInteracaoRequest:
      type: object
      required: [tipo, canal, direcao, conteudo]
      properties:
        tipo:
          type: string
        canal:
          type: string
        direcao:
          type: string
          enum: [ENTRADA, SAIDA, INTERNA]
        conteudo:
          type: string
        visibilidade:
          type: string
          enum: [INTERNA, CIDADAO]
    Interacao:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tipo:
          type: string
        canal:
          type: string
        direcao:
          type: string
          enum: [ENTRADA, SAIDA, INTERNA]
        conteudo:
          type: string
        visibilidade:
          type: string
          enum: [INTERNA, CIDADAO]
        criadaEm:
          type: string
          format: date-time
    HistoricoSolicitacao:
      type: object
      properties:
        id:
          type: string
          format: uuid
        acao:
          type: string
        alteracoes:
          type: object
          additionalProperties: true
        usuarioId:
          type: string
          format: uuid
        criadaEm:
          type: string
          format: date-time
...
    Problem:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string